test(server): split config tests into config_test.go
server/config.go has grown with EffectiveMaxUploadSize and the
backend-specific defaults, so move TestLoadEnvBuckets, TestDefaultConfig,
and TestEffectiveMaxUploadSize out of server_test.go into a new
config_test.go that mirrors the source file layout.
